backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "LLM World Builder Backend"
authors = ["Shawn Chiao <mantle@gmail.com>"]
package-mode = false

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
known_first_party = ["api"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
strawberry-graphql = "^0.219.0"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
alembic = "^1.13.1"
pydantic-settings = "^2.6.1"
anthropic = "^0.40.0"

[tool.poetry.group.dev.dependencies]
black = "^24.1.0"
isort = "^5.13.2"
mypy = "^1.8.0"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
